+++ service/src/main/java/restwars/service/points/PointsDAO.java
@@ -2,6 +2,7 @@ package restwars.service.points;
 
 import restwars.model.points.Points;
 
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -22,5 +23,14 @@ public interface PointsDAO {
      * @param playerId Id of the player.
      * @return Points.
      */
-    Optional<Long> findMostRecentPointsWithPlayerId(UUID playerId);
+    Optional<Points> findMostRecentPointsWithPlayerId(UUID playerId);
+
+    /**
+     * Finds all points for the player with the given id.
+     *
+     * @param playerId Id of the player.
+     * @param max      Maximum number of points.
+     * @return Points.
+     */
+    List<Points> findPointsWithPlayerId(UUID playerId, int max);
 }